requirements.txt
requests==2.32.5
python-telegram-bot[job-queue]==22.6
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2024.1
